refactor: split HR setting panes into components
