+++ client/tests/hrManagementSystemSetting.spec.js
@@ -2,6 +2,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { mount } from '@vue/test-utils'
 import ElementPlus from 'element-plus'
 import HRManagementSystemSetting from '../src/components/backComponents/HRManagementSystemSetting.vue'
+import EmployeeManagement from '../src/components/backComponents/EmployeeManagement.vue'
 
 vi.mock('../src/api', () => ({
   apiFetch: vi.fn(() => Promise.resolve({ ok: true, json: async () => [] }))
@@ -20,19 +21,21 @@ describe('HRManagementSystemSetting.vue', () => {
     const wrapper = mount(HRManagementSystemSetting, {
       global: { plugins: [ElementPlus] }
     })
-    expect(wrapper.vm.employeeDialogVisible).toBe(false)
+    const emp = wrapper.findComponent(EmployeeManagement)
+    expect(emp.vm.employeeDialogVisible).toBe(false)
     const button = wrapper.findAll('button').find(b => b.text() === '新增員工')
     await button.trigger('click')
-    expect(wrapper.vm.employeeDialogVisible).toBe(true)
+    expect(emp.vm.employeeDialogVisible).toBe(true)
   })
 
   it('adds and removes experience rows', async () => {
     const wrapper = mount(HRManagementSystemSetting, {
       global: { plugins: [ElementPlus] }
     })
-    wrapper.vm.addExperience()
-    expect(wrapper.vm.employeeForm.experiences.length).toBe(1)
-    wrapper.vm.removeExperience(0)
-    expect(wrapper.vm.employeeForm.experiences.length).toBe(0)
+    const emp = wrapper.findComponent(EmployeeManagement)
+    emp.vm.addExperience()
+    expect(emp.vm.employeeForm.experiences.length).toBe(1)
+    emp.vm.removeExperience(0)
+    expect(emp.vm.employeeForm.experiences.length).toBe(0)
   })
 })
